reflection-demo/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
google-cloud-secret-manager
python-dotenv
google-generativeai
jsonschema